fix(matador.josh): move login and signup into auth
