+++ Mobile/starter_project_matador/lib/features/auth/domain/repositories/login_repository.dart
@@ -1,0 +1,8 @@
+import 'package:dartz/dartz.dart';
+
+import '../../../../core/error/failures.dart';
+import '../entities/user.dart';
+
+abstract class LoginUserRepository {
+  Future<Either<Failure, AuthUser>> authenticate(String email, String password);
+}
