+++ Mobile/starter_project_matador/lib/features/auth/domain/repositories/signup_repository.dart
@@ -1,0 +1,8 @@
+import 'package:dartz/dartz.dart';
+import 'package:matador/core/error/failures.dart';
+
+import '../entities/user.dart';
+
+abstract class SignUpRepository {
+  Future<Either<Failure, AuthUser>> register(String email, String password);
+}
