+++ Mobile/starter_project_matador/lib/features/auth/presentation/bloc/login_event.dart
@@ -1,0 +1,7 @@
+abstract class LoginEvent {}
+
+class LoginButtonPressedEvent extends LoginEvent {
+  final String email;
+  final String password;
+  LoginButtonPressedEvent({required this.email, required this.password});
+}
