+++ Mobile/starter_project_matador/lib/features/auth/presentation/widgets/login/logo.dart
