+++ Mobile/starter_project_matador/lib/features/auth/presentation/widgets/signup/signup_logo.dart
