--- Mobile/starter_project_matador/lib/features/login/Domain/entities/user.dart
@@ -1,10 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-class User extends Equatable {
-  final String email;
-  final String password;
-  const User({required this.email, required this.password});
-
-  @override
-  List<Object?> get props => ([email, password]);
-}
