--- Mobile/starter_project_matador/lib/features/login/Domain/repositories/user_repository.dart
@@ -1,8 +1,0 @@
-import 'package:dartz/dartz.dart';
-
-import '../../../../core/error/failures.dart';
-import '../entities/user.dart';
-
-abstract class UserRepository {
-  Future<Either<Failure, User>> authenticate(String email, String password);
-}
