--- Mobile/starter_project_matador/lib/features/login/Presentation/widgets/sign_in_with_your_account.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../../../core/utils/constants/global_variables.dart';
-
-class SignInWithYourAccount extends StatelessWidget {
-  final double width;
-
-  const SignInWithYourAccount({Key? key, required this.width})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: width,
-      child: const Text(
-        'Sign in with your account',
-        style: TextStyle(
-            color: secondaryColor,
-            fontWeight: FontWeight.bold,
-            fontFamily: 'Poppins',
-            fontSize: 15),
-      ),
-    );
-  }
-}
