--- Mobile/starter_project_matador/lib/features/login/data/datasources/login_local_datasource.dart
@@ -1,7 +1,0 @@
-import 'package:matador/features/login/Domain/entities/user.dart';
-
-import '../models/login_model.dart';
-
-abstract class LoginUserLocalDataSource {
-  Future<LoginModel> authenticate(User user);
-}
