--- Mobile/starter_project_matador/lib/features/login/data/network_info.dart
@@ -1,3 +1,0 @@
-abstract class NetworkInfo {
-  Future<bool> get isConnected;
-}
