--- Mobile/starter_project_matador/lib/features/signup/domain/entities/user.dart
@@ -1,15 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
-
-class User extends Equatable {
-  final String email;
-  final String password;
-
-  User({
-    required this.email,
-    required this.password,
-  });
-
-  @override
-  List<Object> get props => [email, password];
-}
