--- Mobile/starter_project_matador/lib/features/signup/domain/repositories/auth_repository.dart
@@ -1,8 +1,0 @@
-import 'package:dartz/dartz.dart';
-import 'package:matador/core/error/failures.dart';
-import 'package:matador/features/signup/domain/entities/user.dart';
-
-
-abstract class AuthRepository {
-  Future<Either<Failure, User>> signUp(String email, String password);
-}
