+++ Mobile/starter_project_matador/test/features/auth/data/models/login_model_test.dart
@@ -1,5 +1,5 @@
-import 'package:matador/features/login/Domain/entities/user.dart';
-import 'package:matador/features/login/data/models/login_model.dart';
+import 'package:matador/features/auth/data/models/login_model.dart';
+import 'package:matador/features/auth/domain/entities/user.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:dartz/dartz.dart';
@@ -11,7 +11,7 @@ void main() {
   test(
     'should be a subclass of User entity',
     () async {
-      expect(loginModel, isA<User>());
+      expect(loginModel, isA<AuthUser>());
     },
   );
 }
