+++ Mobile/starter_project_matador/test/features/auth/data/models/login_model_test.mocks.dart
@@ -3,7 +3,7 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:matador/features/login/data/models/login_model.dart' as _i2;
+import 'package:matador/features/auth/data/models/login_model.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
