+++ Mobile/starter_project_matador/test/features/auth/domain/usecases/authenticate_user_test.dart
@@ -1,35 +1,36 @@
 import 'package:dartz/dartz.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:matador/core/usecases/usecases.dart';
-import 'package:matador/features/login/Domain/entities/user.dart';
-import 'package:matador/features/login/Domain/repositories/user_repository.dart';
-import 'package:matador/features/login/Domain/usecases/login_use_case.dart';
+import 'package:matador/features/auth/domain/entities/user.dart';
+import 'package:matador/features/auth/domain/repositories/login_repository.dart';
+import 'package:matador/features/auth/domain/usecases/login_use_case.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
+import '../../../user/domain/usecases/get_user_by_id_test.mocks.dart';
 import 'authenticate_user_test.mocks.dart';
 
-@GenerateMocks([UserRepository])
+@GenerateMocks([LoginUserRepository])
 void main() {
-  GetUser? usecase;
-  MockUserRepository? mockUserRepository;
+  LoginUserUseCase? usecase;
+  MockLoginUserRepository? mockLoginUserRepository;
   setUp(() {
-    mockUserRepository = MockUserRepository();
-    usecase = GetUser(mockUserRepository);
+    mockLoginUserRepository = MockLoginUserRepository();
+    usecase = LoginUserUseCase(mockLoginUserRepository);
   });
   const email = 'test@gmail.com';
   const password = 'password';
-  final tuser = User(email: 'test@gmail.com', password: 'password');
+  final tuser = AuthUser(email: 'test@gmail.com', password: 'password');
   test('should get the user from the user repository', () async {
     //arrange
-    when(mockUserRepository!.authenticate('test@gmail.com', 'password'))
+    when(mockLoginUserRepository!.authenticate('test@gmail.com', 'password'))
         .thenAnswer((_) async => Right(tuser));
     //act
     final result =
         await usecase!(const AuthParams(email: email, password: password));
     //assert
     expect(result, Right(tuser));
-    verify(mockUserRepository!.authenticate(email, password));
-    verifyNoMoreInteractions(mockUserRepository);
+    verify(mockLoginUserRepository!.authenticate(email, password));
+    verifyNoMoreInteractions(mockLoginUserRepository);
   });
 }
