+++ Mobile/starter_project_matador/test/features/auth/domain/usecases/signup_user_test.mocks.dart
@@ -1,14 +1,15 @@
 // Mocks generated by Mockito 5.4.0 from annotations
-// in matador/test/features/auth/data/repositories/login_repository_test.dart.
+// in matador/test/features/auth/domain/usecases/signup_user_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i4;
 
-import 'package:matador/features/auth/data/datasources/login_remote_datasource.dart'
+import 'package:dartz/dartz.dart' as _i2;
+import 'package:matador/core/error/failures.dart' as _i5;
+import 'package:matador/features/auth/domain/entities/user.dart' as _i6;
+import 'package:matador/features/auth/domain/repositories/signup_repository.dart'
     as _i3;
-import 'package:matador/features/auth/data/models/login_model.dart' as _i2;
-import 'package:matador/features/auth/domain/entities/user.dart' as _i5;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -22,8 +23,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeLoginModel_0 extends _i1.SmartFake implements _i2.LoginModel {
-  _FakeLoginModel_0(
+class _FakeEither_0<L, R> extends _i1.SmartFake implements _i2.Either<L, R> {
+  _FakeEither_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -32,28 +33,37 @@ class _FakeLoginModel_0 extends _i1.SmartFake implements _i2.LoginModel {
         );
 }
 
-/// A class which mocks [LoginUserRemoteDataSource].
+/// A class which mocks [SignUpRepository].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockLoginUserRemoteDataSource extends _i1.Mock
-    implements _i3.LoginUserRemoteDataSource {
-  MockLoginUserRemoteDataSource() {
+class MockSignUpRepository extends _i1.Mock implements _i3.SignUpRepository {
+  MockSignUpRepository() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i4.Future<_i2.LoginModel> authenticate(_i5.AuthUser? user) =>
+  _i4.Future<_i2.Either<_i5.Failure, _i6.AuthUser>> register(
+    String? email,
+    String? password,
+  ) =>
       (super.noSuchMethod(
         Invocation.method(
-          #authenticate,
-          [user],
+          #register,
+          [
+            email,
+            password,
+          ],
         ),
-        returnValue: _i4.Future<_i2.LoginModel>.value(_FakeLoginModel_0(
+        returnValue: _i4.Future<_i2.Either<_i5.Failure, _i6.AuthUser>>.value(
+            _FakeEither_0<_i5.Failure, _i6.AuthUser>(
           this,
           Invocation.method(
-            #authenticate,
-            [user],
+            #register,
+            [
+              email,
+              password,
+            ],
           ),
         )),
-      ) as _i4.Future<_i2.LoginModel>);
+      ) as _i4.Future<_i2.Either<_i5.Failure, _i6.AuthUser>>);
 }
